+++ app/src/main/java/com/edu/kotlindbnetwork/AppComponent.kt
@@ -1,7 +1,7 @@
 package com.edu.kotlindbnetwork;
 
-import com.edu.kotlindbnetwork.db.Database;
-import com.edu.kotlindbnetwork.db.network.APIService
+import com.edu.kotlindbnetwork.data.db.Database;
+import com.edu.kotlindbnetwork.data.network.APIService
 import com.edu.kotlindbnetwork.modules.RepoModule
 import com.edu.kotlindbnetwork.repo.UserRepo
 import com.edu.kotlindbnetwork.ui.MainActivity
