+++ app/src/main/java/com/edu/kotlindbnetwork/data/db/Database.kt
@@ -1,9 +1,9 @@
-package com.edu.kotlindbnetwork.db
+package com.edu.kotlindbnetwork.data.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.edu.kotlindbnetwork.db.data.user.User
-import com.edu.kotlindbnetwork.db.data.user.UserDao
+import com.edu.kotlindbnetwork.data.db.user.User
+import com.edu.kotlindbnetwork.data.db.user.UserDao
 
 
 @Database(entities = [User::class], version = 1)
