+++ app/src/main/java/com/edu/kotlindbnetwork/data/db/user/UserDao.kt
@@ -1,4 +1,4 @@
-package com.edu.kotlindbnetwork.db.data.user
+package com.edu.kotlindbnetwork.data.db.user
 
 import androidx.room.*
