+++ app/src/main/java/com/edu/kotlindbnetwork/repo/UserRepoDecorator.kt
@@ -1,7 +1,7 @@
 package com.edu.kotlindbnetwork.repo
 
 import com.edu.kotlindbnetwork.Consts
-import com.edu.kotlindbnetwork.db.data.user.User
+import com.edu.kotlindbnetwork.data.db.user.User
 import java.lang.Exception
 
 class UserRepoDecorator(
